Delete History, Delete Owner and Edit Owner 21-23

+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -116,7 +116,6 @@ namespace MyVet.Web.Controllers
             return View(petType);
         }
 
-        // GET: PetTypes/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -125,26 +124,24 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .Include(pt => pt.Pets)
+                .FirstOrDefaultAsync(pt => pt.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
+            if(petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty,"The pet type can't be removed.");
+                return RedirectToAction(nameof(Index));
+            }
 
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
-
+               
         private bool PetTypeExists(int id)
         {
             return _context.PetTypes.Any(e => e.Id == id);
